Skip request validation for now due to mockbin.com/har shutdown
This used to return some kind of test HAR file (presumably of the
incoming request) but it now returns a cloudflare block page, seemingly
due to changes at Insomnia to remove this.

We can't do much about that in the very short term, so I've disabled
those tests. We should update the test output to use a different
testserver some time soon.