Commit all-escapes test case with all working (apparently) safe outputs
Goal here is that for all _valid_ HTTP inputs, we should correctly
escape special characters throughout. This is tricky to get 100% right
in so many different languages and targets, but this gives us a test
case and a set of checked correctly outputs, and we can incrementally
fix the other cases with broken escaping to resolve the remaining cases
until everything is escaped everywhere correctly.