Preserve non-standard/unparseable query strings as-is in snippet output
This is particularly notable for cases like the &&& repro here, but is
also clearly visible in cases like ?a=b&c, which would often become
something more like '?a=b&c='.

This only applies when the queryString in the HAR is empty, which will
be rare, but HTTP Toolkit (and others) can use this to leave that blank
and send only the raw URL to preserve formatting at the cost of the
losing the nice structure query param setting code in the output.