Improve handling of partially parseable query strings
Previously we assumed that if qs.stringify didn't match the query
string, then the queryString & queryObj properties would be empty, and
so leaving the querystring in the URL (basically disabling structured
generation of query = { a: b } style code and using raw URLs instead)
would be safe.

It turns out though that for partially parseable cases, we can get HARs
with a URL containing a broken query string _and_ a valid queryString
property with some values in it. We now actively ignore the latter, so
the URL as always used as-is.